Fix duplicate tpl-onavg dataset
The Canadian Open Neuroscience Platform (CONP) Portal is a web interface that facilitates open science for the neuroscience community by simplifying global access to and sharing of datasets and tools.

One of our dataset uses templateflow as a submodule dependency. Datalad fails to install it because of the multiplicity of the tpl-onavg module. We suspect this duplication is a typo and can be removed, there we would like to propose the change.


We built a portal, CONP (https://portal-dev.conp.ca/) We use Datalad into install  Fix duplicate dataset that causes datalad to fail